Use enqueue for css/script assets (#8)
According to WP, we should load assets via

https://developer.wordpress.org/reference/functions/wp_enqueue_script/
https://developer.wordpress.org/reference/functions/wp_enqueue_style/
https://developer.wordpress.org/reference/functions/wp_add_inline_script/

This is in response to feedback from the Wordpress plugin submission. We
register the assets and then actually enqueue them when a shortcode is
called.

Note: this is different from the normal integration flow of async/defer.
This is because the script management system of WP makes it difficult to
run a script given our Answers dependencies.

Notable changes: the answers sdk script call is no longer async/defer.
The initAnswers call is called after the script is loaded, not in the
onload.

TEST=manual

Test that the scripts are not loaded if the shortcode is not called.
Test that the [yext_searchbar] and [yext_results_page] still work.

Make sure the CSS is loading correctly and the inline JS calling
initAnswers() is /after/ the call to the synchronous answers.js

Test on a test page in the Yext WP instance.
Test on a test page in a local WP instance.