Update Tested Up To after testing on 5.6 WordPress (#11)
Change readme to reflect that plugin was tested on 5.6. After this is merged, will
move the changes to the SVN repository, create a 1.0.1 folder in the tags/ folder
with the new code (readme changes).

J=SLAP-934
TEST=manual

Test that do_shortcode, shortcode block for [yext_resultspage] and [yext_searchbar]
all work in putting an experience on a page.